requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
starlette==0.48.0
pydantic==2.11.9
pydantic_core==2.33.2
annotated-types==0.7.0

python-dotenv==1.1.1
PyJWT==2.10.1
bcrypt==5.0.0
cryptography==46.0.2

mariadb==1.1.13
pandas==2.1.4
numpy==1.26.4
openpyxl==3.1.2
python-dateutil==2.9.0.post0
pytz==2025.2

selenium==4.16.0
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1

google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

requests==2.32.5